requirements.txt
mlflow==2.8.1
kafka-python==2.0.2
pandas
numpy
scikit-learn
pycaret==3.2.0
fastapi==0.104.1
uvicorn==0.24.0
prometheus-client==0.19.0
pytest==7.4.3
psycopg2-binary==2.9.9
requests==2.31.0
email-validator>=2.0.0
